Add the TABuffer
This commit has (at least) two hacks:

1. TAs need a region_id and element_id. The region_id is taken from
configuration, but there's only one TA buffer per region, so
element_id doesn't really apply in the current scheme. For now I've
set the element_id to 10,000 for the TA buffers, which will at least
avoid collisions with the TP buffers. The resulting HDF5 files look
like:

      GROUP "Trigger" {
         GROUP "Region000" {
            DATASET "Element00" {
               DATATYPE  H5T_STD_I8LE
               DATASPACE  SIMPLE { ( 248, 1 ) / ( 248, 1 ) }
            }
            DATASET "Element01" {
               DATATYPE  H5T_STD_I8LE
               DATASPACE  SIMPLE { ( 248, 1 ) / ( 248, 1 ) }
            }
            DATASET "Element10000" {
               DATATYPE  H5T_STD_I8LE
               DATASPACE  SIMPLE { ( 184, 1 ) / ( 184, 1 ) }
            }
         }
      }

2. We need to fan-out TAs from the TA maker to the TA buffer and the
TC maker. But there's no DAQ module which will do that for us, so
instead the TA maker sends the TAs to an external pub/sub endpoint, and the TA
buffer and TC maker read the TAs from that endpoint. The connection is
made in daqconf_multiru_gen, rather than in trigger_gen.py, which is
unfortunate, plus we pay for a serialization and 2 deserializations
that could be replaced by some copy ctors if we had a FanoutDAQModule
or equivalent.